requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
python-dotenv
streamlit